+++ actionpack/lib/action_dispatch/railtie.rb
@@ -36,6 +36,7 @@ module ActionDispatch
     config.action_dispatch.debug_exception_log_level = :fatal
     config.action_dispatch.strict_freshness = false
 
+    config.action_dispatch.dom_testing_assertions = :rails_dom_testing
     config.action_dispatch.ignore_leading_brackets = nil
     config.action_dispatch.strict_query_string_separator = nil
     config.action_dispatch.verbose_redirect_logs = false
@@ -102,6 +103,21 @@ module ActionDispatch
 
       ActionDispatch::Http::Cache::Request.strict_freshness = app.config.action_dispatch.strict_freshness
       ActionDispatch.test_app = app
+    end
+
+    initializer "action_dispatch.integration_test" do |app|
+      dom_testing_assertions = app.config.action_dispatch.delete(:dom_testing_assertions)
+
+      ActiveSupport.on_load(:action_dispatch_integration_test) do
+        case dom_testing_assertions
+        when :rails_dom_testing
+          include ActionView::RailsDomTestingAssertions
+        when :none
+          # do nothing
+        else
+          raise ArgumentError.new("unrecognized value #{assertions.inspect} for config.action_dispatch.dom_testing_assertions")
+        end
+      end
     end
   end
 end